+++ src/fields/Users.php
@@ -61,6 +61,7 @@ class Users extends Field implements FieldInterface
     public function parseField(): mixed
     {
         $value = $this->fetchArrayValue();
+        $default = $this->fetchDefaultArrayValue();
 
         $sources = Hash::get($this->field, 'settings.sources');
         $limit = Hash::get($this->field, 'settings.limit');
@@ -89,8 +90,15 @@ class Users extends Field implements FieldInterface
 
         foreach ($value as $dataValue) {
             // Prevent empty or blank values (string or array), which match all elements
-            if (empty($dataValue)) {
+            if (empty($dataValue) && empty($default)) {
                 continue;
+            }
+
+            // special provision for falling back on default BaseRelationField value
+            // https://github.com/craftcms/feed-me/issues/1195
+            if (empty($dataValue) && !empty($default)) {
+                $foundElements = $default;
+                break;
             }
 
             // If we're using the default value - skip, we've already got an id array