+++ src/helpers/DataHelper.php
@@ -10,7 +10,6 @@ use craft\feedme\Plugin;
 use craft\helpers\DateTimeHelper;
 use craft\helpers\Db;
 use craft\helpers\Json;
-use DateTime;
 use Throwable;
 
 class DataHelper
@@ -54,7 +53,6 @@ class DataHelper
         $value = [];
 
         $node = Hash::get($fieldInfo, 'node');
-        $default = Hash::get($fieldInfo, 'default');
 
         $dataDelimiter = Plugin::$plugin->service->getConfig('dataDelimiter');
 
@@ -68,10 +66,6 @@ class DataHelper
             $feedPath = preg_replace('/^(\d+\/)|(\/\d+)/', '', $feedPath);
 
             if ($feedPath == $node || $nodePath == $node) {
-                if ($nodeValue === null || $nodeValue === '') {
-                    $nodeValue = $default;
-                }
-
                 // Allow pipes '|' to denote multiple items, but even if it doesn't contain one, explode will create
                 // an array, so ensure to merge with the current results.
                 if (is_string($nodeValue) && str_contains($nodeValue, $dataDelimiter)) {
@@ -89,13 +83,25 @@ class DataHelper
 
         // Check if not importing, just using default
         if ($node === 'usedefault' && !$value) {
-            if (!is_array($default)) {
-                $default = [$default];
-            }
-            $value = $default;
+            $value = self::fetchDefaultArrayValue($fieldInfo);
         }
 
         return $value;
+    }
+
+    /**
+     * @param $fieldInfo
+     * @return array|\ArrayAccess|mixed
+     */
+    public static function fetchDefaultArrayValue($fieldInfo)
+    {
+        $default = Hash::get($fieldInfo, 'default');
+
+        if (!is_array($default)) {
+            $default = [$default];
+        }
+
+        return $default;
     }
 
     /**